requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs

#ML datasets
datasets


# utilities
six>=1.14
joblib
memoized-property
termcolor

# scraping
requests
httplib2
bs4
Pillow
fake-useragent
selenium


# face detection
facenet-pytorch


# deepl learning
deepface
tensorflow==2.5.2
